Overall structure and solve simplification
There are no substantial changes to functionality in this commit.

Aesthetically, though…

* Solve

The solve has been broken up a little further, and `?`s are now used extensively to clarify the basic procedure.
A few things have also been moved to a more sensible place, …

Some attempt has been made to tidy the gen and errs a little, also.
These are simplfified a little for the solve, and otherwise further concentrated in the types submodule.

* Structure/Context

Various things have been moved around, most not worth noting in detail.

From a high level, though, the basic idea is to treat the context as the central object, and to assume it has impl's throughout the rest of the module.
With this, e.g., it may be expected that methods for representing the valuation as a string via the context are found somewhere amoung other valuation methods.

Another notable change is moving the consequence q to the db submodule.
Intutively, this works as the q is a q of what to check to ensure consistency among the other databases.

+++ otter_lib/src/generic/luby.rs
@@ -2,7 +2,7 @@
 // https://gitlab.com/chaotic-evil/luby
 // https://doc.rust-lang.org/rust-by-example/trait/iter.html
 
-pub type LubyType = u32;
+pub type LubyType = usize;
 
 pub struct Luby {
     curr: LubyType,
